stable 9.6:push dev changes to stable build
61e02cb8 Make the easter_egg script more CPU friendly
4c8378bc fix https://github.com/Ansuel/tch-nginx-gui/issues/990
ccff386e added few more checks and commits to network config code reformat
9e163354 should fix https://github.com/Ansuel/tch-nginx-gui/issues/1001 and https://github.com/Ansuel/tch-nginx-gui/issues/990 set AGTEF gateway image for TG789vac v3
a232c9db Fix missing exec flag for transfer_basefiles
149e529d Abuse the prozone to speedup migration
40847eef Fix migration in 18+ version
a6bbdd63 Fix bad name for transfer_basefiles
bf01554e fix wget for TG800 with upgrade pack remove debris after DGA upgrade pack install
66a6ed25 opkg set feed by major version
d24379a5 fix https://github.com/Ansuel/tch-nginx-gui/issues/986 enable DGA apps/repos to TG800
ed0ff8fc fix typo in migrate_overlay script
aa9ba614 fix migrate_overlay to copy deleted files
9ab56bf3 try to fix config reset on migration
77435abe update TIM cwmp latest version
7de532d6 stop watchdog while migrating
b2dbd707 drop backup from migrate_overlay
b0b529fb fix bad if logic in migrate_overlay
7db4dfa5 try to fix migrate overlay problems
b7c7f327 add wait after issuing reboot after migration to block other initscript
0fabac58 fix https://github.com/Ansuel/tch-nginx-gui/issues/980 fix symlink to telnetd init fix ledfw package removal
fcf3a364 fix https://github.com/Ansuel/tch-nginx-gui/issues/979
b1750b92 add A2pvfbH045r xdsl driver
c4413ee2 add insmod hook fix https://github.com/Ansuel/tch-nginx-gui/issues/976
ea81db1c fix https://github.com/Ansuel/tch-nginx-gui/issues/976 fix modeoverlay hook on <16.2
5593db7a fix https://github.com/Ansuel/tch-nginx-gui/issues/928 simplified background curl for changelog.lp
1cd25881 try fix https://github.com/Ansuel/tch-nginx-gui/issues/963 more fixes to debughelper
a2d331f3 fix and add more infos to DebugHelper
eb318432 xdsl drivers: add A2pvfbH045q
c2164d25 cleanup voice and bridge mode from broadband all cards shown when in bridge mode
cf4651af mapper: fix bug with xdslctl
d8fe800c always run rootdevice when generating debug log fix remove icon in mobile page
dca8dece fix https://github.com/Ansuel/tch-nginx-gui/issues/961
53c0be7b mapper: add NumerOfEntries to ddns
12468f82 mapper: imrpove speed
59ea127e mapper fix bug in xdslctl
da0ae0f8 web: use xdslctl to get Firmware Version
2a9e0d2d mapper: add DslamFirmware to xdslctl
9c5c1e54 mapper: use minimalStats for rate
36fb6a1a fix few huawei USB 3G dongle stuck connecting
ad89692b fix path
96d9117c move do_migrate_overlay activation to mount_modroot folder
007e7a12 fix error 500 when restoring advanced backup
8b6c4845 avoid IspConfigHelper debug if not set and fix path not found during install
9192b5cc upgrade: update preserve_file in platform.sh
1433387e rootdevice: restore use of random_seed hook
e9c52dae mount_root: mount in preint, migrate in init
dca85cba mount_root: disable too verbose output
a0a31f9a mount_root: mount modoverlay in init
bb86e713 mount_root: fix reboot bug
cbc88bfd mount_root: change migrate logic (use ram)
4f92821d minor install fixes
7a153a5e mount_root: fix migration script pt3
84460280 mount_root: fix migration script pt2
2da2d184 mount_root: fix broken migration script
7917bea5 change hook method to allow hooking binaries
22efe916 mount_root: reboot after file migration
3a22e1a4 New translations (#956)
85c8a096 forcely save update GUI package to root
776e2e9d better log messages on mount_root-mod
d05975e3 mount_Root: fix some killing errors
1f5f6452 adding few sync use mtd -r when erasing bank1
b2a064c2 more verbose logging for mount_root-mod
016fa8a6 improve safe_mv
3ba2c4be fix path for first install check, make sure to update package when migrating to bank_2
c563d157 Update Italian language (#952)
bd9ed710 BuildBot: translation: update po file automatically
be0f6cdd Safety rework (#950)
30acbd03 BuildBot: translation: update po file automatically
1903e357 allow setup of wan connection when in bridge mode
b7598611 fix LEDs on old DGA firmwares
7cec8715 ci: fix broken script
07dfdb35 use official openwrt feeds for brcm63xx fix ledfw extraction on TG788
1960ba30 improved dropbear wan config check
d56a43a0 minor fix for 788 install
6f40dad1 fix openwrt restore fix appInstallRemoveUtility.sh typo
dab627c6 fix upgradegui space check
8846d147 fix xupnp install/remove
3519b057 fix syntax
d80ec4e2 fix syntax
083ebd43 ci: fix typo in scripts
fec805e3 improved installation of specific app packages show warning if installing extensions offline
6ce9c6dd ci: support custom branch build version
20ac9b7f blue led while booting on all devices
622dff34 Update StateMachines.lua (blue led) (#733)
30e48a78 realign IT webui core to original formatting minor string improvements
1d9838e9 avoid breaking fastweb SIP
9bb08d0c fix GPON card error
b46deb2c fix https://github.com/Ansuel/tch-nginx-gui/issues/917#issuecomment-611466224
5cf5b162 fix inverted bytes metrics in GPON update IT translation
d80373c1 BuildBot: translation: update po file automatically
feb57baf more fixes for GPON https://github.com/Ansuel/tch-nginx-gui/issues/917
3ed1e0da fix https://github.com/Ansuel/tch-nginx-gui/issues/882 moved misplaced opkg install in upgrade packs
589f256b check for null value if cannot communicate with the SFP
1d153c88 fix xdsl drivers download
66074d67 remove all SFP interface changes during install improve GPON modal/card should fix https://github.com/Ansuel/tch-nginx-gui/issues/917
c2c4c496 moved checkver and getvendor to modgui modal
239bd729 fix mobiled perm
f1e66f73 fix mobiled perm
8a0d8c04 fix https://github.com/Ansuel/tch-nginx-gui/issues/924
34a0d890 Remove non-existent hardcoded distfeed to avoid 404 on opkg update
f792188f silence blacklist reinstall when upgrading
de2d7bab fix offline GUI upgrade fix specific app install fix js error in command-log-read-modal.lp remove hardcoded blacklist update script in upgradegui

testing [ci skip]

+++ decompressed/gui_file/etc/modgui_scripts/04_config.sh
@@ -101,31 +101,23 @@ create_driver_setting() {
 dropbear_config_check() {
   if [ ! "$(uci get -q dropbear.wan)" ]; then
     logger_command "Adding Dropbear wan config"
+    uci add dropbear dropbear >/dev/null
+    uci rename dropbear.@dropbear[-1]=wan
     uci set dropbear.wan=dropbear
     uci set dropbear.wan.Interface='wan'
-    uci set dropbear.wan.RootLogin='1'
-    uci set dropbear.wan.RootPasswordAuth='on' #dropbear root related
-    uci set dropbear.wan.PasswordAuth='on'
     uci set dropbear.wan.Port='22'
     uci set dropbear.wan.enable='0'
+  fi
 
+  uci set dropbear.wan.RootLogin='1'
+  uci set dropbear.wan.RootPasswordAuth='on' #dropbear root related
+  uci set dropbear.wan.PasswordAuth='on'
+
+  if [ "$(uci changes)" ]; then
+    logger_command "Restarting Dropbear SSH Server..."
     uci commit dropbear
-  else
-    if [ "$(uci get -q dropbear.wan.RootLogin)" != "1" ]; then
-      logger_command "Enabling Dropbear wan RootLogin"
-      uci set dropbear.wan.RootLogin='1'
-      uci commit dropbear
-    fi
-    if [ "$(uci get -q dropbear.wan.PasswordAuth)" != "on" ]; then #TELMEX firmware got it OFF
-      logger_command "Enabling Dropbear wan PasswordAuth"
-      uci set dropbear.wan.PasswordAuth='on'
-      uci commit dropbear
-    fi
-    if [ "$(uci get -q dropbear.wan.RootPasswordAuth)" != "on" ]; then #TELMEX firmware got it OFF
-      logger_command "Enabling Dropbear wan RootPasswordAuth"
-      uci set dropbear.wan.RootPasswordAuth='on'
-      uci commit dropbear
-    fi
+    /etc/init.d/dropbear enable
+    /etc/init.d/dropbear restart >/dev/null
   fi
 }
 
@@ -136,7 +128,7 @@ eco_param() {
     uci set power.cpu.cpuspeed='256'
     uci set power.cpu.wait='1'
     logger_command "Restarting power management"
-    /etc/init.d/power restart
+    /etc/init.d/power restart &>/dev/null
   fi
 }
 
@@ -186,7 +178,8 @@ create_gui_type() {
   elif [ "$(uci get -q modgui.app.blacklist_app)" = "1" ] &&
     [ ! -f /www/docroot/modals/mmpbx-contacts-modal.lp.orig ] &&
     [ -f /usr/share/transformer/scripts/appInstallRemoveUtility.sh ]; then
-    /usr/share/transformer/scripts/appInstallRemoveUtility.sh install blacklist
+    logger_command "Reinstalling blacklist app after upgrade..."
+    /usr/share/transformer/scripts/appInstallRemoveUtility.sh install blacklist >/dev/null
   fi
 }
 
@@ -233,7 +226,12 @@ suppress_excessive_logging() {
 
 real_ver_entitied() {
   if [ -f /rom/etc/uci-defaults/tch_5000_versioncusto ] && [ -f /etc/config/versioncusto ]; then
-    short_ver="$(grep </proc/banktable/activeversion -Eo '.*\..*\.[0-9]*-[0-9]*')"
+    bank_version="activeversion"
+    if [ "$(cat /proc/banktable/booted)" != "$(cat /proc/banktable/active)" ]; then
+      bank_version="passiveversion"
+    fi
+
+    short_ver="$(grep </proc/banktable/$bank_version -Eo '.*\..*\.[0-9]*-[0-9]*')"
     real_ver=$(grep </rom/etc/uci-defaults/tch_5000_versioncusto "$short_ver" | awk '{print $2}')
     uci set versioncusto.override.fwversion_override_latest="$latest_version_on_TIM_cwmp"
     if [ "$real_ver" = "" ]; then
@@ -296,13 +294,6 @@ add_xdsl_option() {
   uci set xdsl.dsl0.demod_cap_value="0x00047a"
 }
 
-check_wan_mode() {
-  if [ ! "$(uci get -q network.config.wan_mode)" ]; then
-    uci set network.config="config"
-    uci set network.config.wan_mode="$(uci get -q network.wan.proto)"
-  fi
-}
-
 dosprotect_inizialize() {
   logger_command "Checking DoSprotect kernel modules..."
   if [ -n "$(find /lib/modules/*/ -iname xt_hashlimit.ko)" ]; then
@@ -340,7 +331,7 @@ mobiled_lib_add() {
     /etc/init.d/mobiled restart
   else
     #make sure we haven't replaced it some old GUI install, restore from rom if needed
-    if [ "$(md5sum /rom/etc/init.d/mobiled | cut -d' ' -f1)" != "$(md5sum /etc/init.d/mobiled | cut -d' ' -f1)" ]; then
+    if [[ -f /rom/etc/init.d/mobiled && -n "$(cmp /rom/etc/init.d/mobiled /etc/init.d/mobiled)" ]]; then
       logger_command "Restoring and restarting /etc/init.d/mobiled ..."
       cp /rom/etc/init.d/mobiled /etc/init.d/mobiled
       /etc/init.d/mobiled restart
@@ -361,8 +352,8 @@ mobiled_lib_add() {
   major_system_version="$(uci get version.@version[0].marketing_version | sed 's#\.##' | grep -o -E '[0-9]+')"
   if [ "$major_system_version" -lt 173 ]; then #if fw <17.3
     #Restore original lte-doctor related webui files
-    cp /rom/www/docroot/ajax/radioparameters.lua /www/docroot/ajax/radioparameters.lua
-    cp /rom/www/docroot/modals/lte-doctor.lp /www/docroot/modals/lte-doctor.lp
+    [ -f /rom/www/docroot/ajax/radioparameters.lua ] && cp /rom/www/docroot/ajax/radioparameters.lua /www/docroot/ajax/radioparameters.lua
+    [ -f /rom/www/docroot/modals/lte-doctor.lp ] && cp /rom/www/docroot/modals/lte-doctor.lp /www/docroot/modals/lte-doctor.lp
   fi
 }
 
@@ -476,7 +467,7 @@ cumulative_check_gui() {
     logger_command "Can't generate GUI hash, file not found!"
     gui_hash="0"
   fi
-  
+
   saved_gui_version=$(uci get -q modgui.gui.gui_version | cut -d'-' -f1)
   clean_version_gui=$(echo "$version_gui" | cut -d'-' -f1)
 
@@ -521,13 +512,13 @@ cumulative_check_gui() {
   if [ ! "$(uci get -q modgui.gui.gui_animation)" ]; then
     uci set modgui.gui.gui_animation="1"
   fi
-  if [ -z "$(uci get -q modgui.var.isp_autodetect)" ]; then
+  if [ ! "$(uci get -q modgui.var.isp_autodetect)" ]; then
   	uci set modgui.var.isp_autodetect="1"
   fi
-  if [ -z "$(uci get -q modgui.var.isp_autodetect)" ]; then
+  if [ ! "$(uci get -q modgui.var.isp)" ]; then
   	uci set modgui.var.isp="Other"
   fi
-  if [ -z "$(uci get -q modgui.gui.gui_skin)" ]; then
+  if [ ! "$(uci get -q modgui.gui.gui_skin)" ]; then
     uci set modgui.gui.gui_skin="green"
   fi
 }
@@ -608,7 +599,7 @@ logger_command "Check Dropbear config file"
 dropbear_config_check #check dropbear config
 logger_command "Check eco paramaters"
 eco_param #This disable eco param as they introduce some latency
-logger_command "Create GUI type in config"
+logger_command "Add app-extension var in modgui uci config"
 create_gui_type #Gui Type
 logger_command "Add new web options"
 add_new_web_rule #This check new option so that we don't replace the one present
@@ -625,8 +616,6 @@ logger_command "Apply new xDSL options"
 add_xdsl_option #New xdsl option
 logger_command "Adding fast cache options"
 fcctlsettings_daemon #Adds fast cache options
-logger_command "Checking if wan_mode option exists..."
-check_wan_mode        # wan_mode check
 dosprotect_inizialize #dosprotected inizialize function
 logger_command "Checking mobiled libs..."
 mobiled_lib_add
